+++ lib/notifications/client/response_notification.rb
@@ -7,7 +7,6 @@ module Notifications
       content
       template
       uri
-      postage
     ).freeze
 
       attr_reader(*FIELDS)